+++ theHandsome/src/main/java/com/thehandsome/service/MemberServiceImpl.java
@@ -50,5 +50,9 @@ public class MemberServiceImpl implements MemberService {
 	public int memberWithdrawal(MemberVO member) throws Exception {
 		return mapper.memberWithdrawal(member);
 	}
-
+	//회원 가입시 권한 부여
+	@Override
+	public void memberAuthJoin(String mid) throws Exception {
+		mapper.memberAuthJoin(mid);
+	}
 }// end class
