+++ theHandsome/src/test/java/com/theHandsome/mapper/MakeCouponTest.java
@@ -1,4 +1,4 @@
-package com.theHandsome.persistence;
+package com.theHandsome.mapper;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,26 +7,31 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import com.thehandsome.domain.CouponVO;
 import com.thehandsome.mapper.CouponMapper;
-import lombok.extern.log4j.Log4j;
 
 /*****************************************************
- * @function : GetCouponTest
+ * @function : MakeCouponTest
  * @author : 김민선
  * @Date : 2022.10.21
  *****************************************************/
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration("file:src/main/webapp/WEB-INF/spring/root-context.xml")
-@Log4j
 
-public class GetCouponTest {
+public class MakeCouponTest {
+
 	@Autowired
 	private CouponMapper mapper;
-	// 쿠폰 조회 테스트
+	
+	// 쿠폰 발급 테스트
 	@Test
-	public void getCoupon() throws Exception{
-		CouponVO coupon = mapper.getCoupon("ms");
-		log.info(coupon);
+	public void makeCoupon() throws Exception{
+		CouponVO coupon = new CouponVO();
+		coupon.setCcode("ABC1236");
+		coupon.setMid("ms");
+		coupon.setCname("ms coupon");
+	    mapper.makeCoupon(coupon);
 
 	}
 }
+
+
