+++ theHandsome/src/test/java/com/theHandsome/mapper/MemberLoginTest.java
@@ -1,4 +1,4 @@
-package com.theHandsome.persistence;
+package com.theHandsome.mapper;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -6,37 +6,31 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import com.thehandsome.domain.MemberVO;
-import com.thehandsome.service.MemberService;
+import com.thehandsome.mapper.MemberMapper;
+import lombok.Setter;
+import lombok.extern.log4j.Log4j;
 
 /*****************************************************
- * @function : MemberJoinTest
- * @author : 김민선
+ * @function : MemberLoginTest
+ * @author : 심지연
  * @Date : 2022.10.18
  *****************************************************/
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration("file:src/main/webapp/WEB-INF/spring/root-context.xml")
-
-public class MemberJoinTest {
+@Log4j
+public class MemberLoginTest {
 	
-	@Autowired
-	private MemberService service;
+	@Setter(onMethod_ = @Autowired)
+	private MemberMapper mapper;
 	
-	//회원가입 테스트
 	@Test
-	public void memberJoin() throws Exception{
+	public void memberLogin() throws Exception{
 		MemberVO member = new MemberVO();
 		
-		member.setMid("test5");
-		member.setMpassword("pw03");
-		member.setMemail("test04@naver.com");
-		member.setMname("테스트5");
-		member.setMtel("0101234333");
-		member.setMzipcode("12200");
-		member.setMaddress1("테스트 주소3");
-		member.setMaddress2("테스트 주소3");
-		
-		service.memberJoin(member);
+		member.setMid("test01");
+		member.setMpassword("pw01");
+		mapper.memberLogin(member);
+		System.out.println("결과 값 : " + mapper.memberLogin(member));
 	}
-
-} 
+}
