+++ theHandsome/src/test/java/com/theHandsome/mapper/ProductMapperTest2.java
@@ -1,4 +1,4 @@
-package com.theHandsome.persistence;
+package com.theHandsome.mapper;
 
 import java.util.List;
 import org.junit.Test;
